LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-no-accept-header-expected.txt

CONSOLE MESSAGE: line 0: Error during WebSocket handshake: 'Sec-WebSocket-Accept' header is missing
Test whether WebSocket handshake fails if the server does not send Sec-WebSocket-Accept header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
